organize code based by folders
