+++ src/components/UI/Card.js
@@ -1,4 +1,3 @@
-
 import './Card.css'
 
 function Card(props) {
